refactor: model file architecture
