+++ src/etl/factories/RiphCtisResearchStudyModelFactory.ts
@@ -1,17 +1,17 @@
 import { Extension, Identifier, Meta } from 'fhir/r4'
 
-import { ModelUtils } from '../../shared/models/custom/ModelUtils'
-import { ReferenceContentsModel } from '../../shared/models/custom/ReferenceContentsModel'
-import { CodeableConceptModel } from '../../shared/models/fhir/DataTypes/CodeableConceptModel'
-import { IdentifierModel } from '../../shared/models/fhir/DataTypes/IdentifierModel'
-import { GroupModel } from '../../shared/models/fhir/GroupModel'
-import { LocationModel } from '../../shared/models/fhir/LocationModel'
-import { ContactDetailModel } from '../../shared/models/fhir/MetadataTypes/ContactDetailModel'
-import { OrganizationModel } from '../../shared/models/fhir/OrganizationModel'
-import { RiphStatus, ResearchStudyModel } from '../../shared/models/fhir/ResearchStudyModel'
-import { ExtensionModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/ExtensionModel'
-import { MetaModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/MetaModel'
-import { ReferenceModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/ReferenceModel'
+import { CodeableConceptModel } from '../../shared/models/data-types/CodeableConceptModel'
+import { IdentifierModel } from '../../shared/models/data-types/IdentifierModel'
+import { GroupModel } from '../../shared/models/domain-resources/GroupModel'
+import { LocationModel } from '../../shared/models/domain-resources/LocationModel'
+import { OrganizationModel } from '../../shared/models/domain-resources/OrganizationModel'
+import { RiphStatus, ResearchStudyModel } from '../../shared/models/domain-resources/ResearchStudyModel'
+import { ModelUtils } from '../../shared/models/eclaire/ModelUtils'
+import { ReferenceContentsModel } from '../../shared/models/eclaire/ReferenceContentsModel'
+import { ContactDetailModel } from '../../shared/models/metadata-types/ContactDetailModel'
+import { ExtensionModel } from '../../shared/models/special-purpose-data-types/ExtensionModel'
+import { MetaModel } from '../../shared/models/special-purpose-data-types/MetaModel'
+import { ReferenceModel } from '../../shared/models/special-purpose-data-types/ReferenceModel'
 import { RiphCtisDto } from '../dto/RiphCtisDto'
 
 export class RiphCtisResearchStudyModelFactory {
