+++ src/etl/factories/RiphJardeResearchStudyModelFactory.ts
@@ -1,16 +1,16 @@
 import { Extension, Identifier, Organization } from 'fhir/r4'
 
-import { ModelUtils } from '../../shared/models/custom/ModelUtils'
-import { ReferenceContentsModel } from '../../shared/models/custom/ReferenceContentsModel'
-import { CodeableConceptModel } from '../../shared/models/fhir/DataTypes/CodeableConceptModel'
-import { IdentifierModel } from '../../shared/models/fhir/DataTypes/IdentifierModel'
-import { GroupModel } from '../../shared/models/fhir/GroupModel'
-import { ContactDetailModel } from '../../shared/models/fhir/MetadataTypes/ContactDetailModel'
-import { OrganizationModel } from '../../shared/models/fhir/OrganizationModel'
-import { ResearchStudyModel, RiphStatus } from '../../shared/models/fhir/ResearchStudyModel'
-import { ExtensionModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/ExtensionModel'
-import { MetaModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/MetaModel'
-import { ReferenceModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/ReferenceModel'
+import { CodeableConceptModel } from '../../shared/models/data-types/CodeableConceptModel'
+import { IdentifierModel } from '../../shared/models/data-types/IdentifierModel'
+import { GroupModel } from '../../shared/models/domain-resources/GroupModel'
+import { OrganizationModel } from '../../shared/models/domain-resources/OrganizationModel'
+import { ResearchStudyModel, RiphStatus } from '../../shared/models/domain-resources/ResearchStudyModel'
+import { ModelUtils } from '../../shared/models/eclaire/ModelUtils'
+import { ReferenceContentsModel } from '../../shared/models/eclaire/ReferenceContentsModel'
+import { ContactDetailModel } from '../../shared/models/metadata-types/ContactDetailModel'
+import { ExtensionModel } from '../../shared/models/special-purpose-data-types/ExtensionModel'
+import { MetaModel } from '../../shared/models/special-purpose-data-types/MetaModel'
+import { ReferenceModel } from '../../shared/models/special-purpose-data-types/ReferenceModel'
 import { RiphJardeDto } from '../dto/RiphJardeDto'
 
 export class RiphJardeResearchStudyModelFactory {
