+++ src/etl/ingest-pipelines/IngestPipeline.ts
@@ -2,7 +2,7 @@ import { errors } from '@elastic/elasticsearch'
 
 import { ElasticsearchService } from '../../shared/elasticsearch/ElasticsearchService'
 import { LoggerService } from '../../shared/logger/LoggerService'
-import { ResearchStudyModel } from '../../shared/models/fhir/ResearchStudyModel'
+import { ResearchStudyModel } from '../../shared/models/domain-resources/ResearchStudyModel'
 import { RiphCtisDto } from '../dto/RiphCtisDto'
 import { RiphDmDto } from '../dto/RiphDmDto'
 import { RiphJardeDto } from '../dto/RiphJardeDto'
