+++ src/shared/models/data-types/PeriodModel.ts
