+++ src/shared/models/data-types/QuantityModel.ts
