+++ src/shared/models/data-types/RangeModel.test.ts
