+++ src/shared/models/data-types/RangeModel.ts
