+++ src/shared/models/eclaire/ReferenceContentsModel.ts
