+++ src/shared/models/special-purpose-data-types/ReferenceModel.ts
